+++ changelog.md
@@ -1,10 +1,11 @@
 # Version 2024.11.5 (2024-11-19)
 
 - Fix returned version of client
+- Improve store tmp value
 
 # Version 2024.11.4 (2024-11-19)
 
-- - Add sysvar/program refresh to scheduler
+- Add sysvar/program refresh to scheduler
 - Run periodic tasks with an individual interval
 - Store tmp value for polling client data points